SCRUM-207 refactor: move payment flow helpers from lib to model
