+++ src/features/subscriptions/model/index.ts
@@ -1,4 +1,7 @@
 export * from './mapSubscriptionTypeToLabel'
 export * from './mapPaymentTypeToLabel'
 export * from './payments.constants'
+export * from './paymentBaseline'
+export * from './paymentPolling'
+export * from './paymentPending'
 export * from './payments.types'
